Fixed a nasty bug in the calculateBias function. Might be the reason why the RNNs are not working for me.
Fixed a nasty bug in the calculateBias function. Might be the reason why
the RNNs are not working for me.

Found it by contrasting the gradients of my own reverse mode AD with
DiffSharp's. Hopefully by the time I am done with this, I'll figure out
why the logistic cross entropy cost function is not working with
DiffSharp.